Add floatValue() to amounts because it's needed in graphs

+++ ui-compose/src/commonMain/kotlin/org/hildan/accounting/ui/plots/AmountAxisModel.kt
@@ -26,5 +26,5 @@ fun amountAxisModel(
 
 private object AmountFloatConverter : FloatConverter<Amount> {
     override fun convertTo(value: Float): Amount = Amount(value.toString())
-    override fun convertFrom(value: Amount): Float = value.doubleValue().toFloat()
+    override fun convertFrom(value: Amount): Float = value.floatValue()
 }
